+++ app/src/main/java/com/kgeorgiev/stackoverflowusers/domain/model/User.kt
@@ -1,7 +1,9 @@
 package com.kgeorgiev.stackoverflowusers.domain.model
 
-data class User(val accountId:Int,
-                val displayName:String,
-                val reputation:Int,
-                val profileImageUrl:String,
-                val isFollowed: Boolean = false)
+data class User(
+    val accountId: Long,
+    val displayName: String,
+    val reputation: Int,
+    val profileImageUrl: String,
+    val isFollowed: Boolean = false
+)
